[#102] refactor : LogList 컴포넌트 공용에서 개별로 이동
